Make overview dynamic based on the indicator type
