+++ x-pack/plugins/observability/public/pages/slo_details/components/overview/overview.stories.tsx
@@ -8,8 +8,8 @@
 import React from 'react';
 import { ComponentStory } from '@storybook/react';
 
-import { KibanaReactStorybookDecorator } from '../../../utils/kibana_react.storybook_decorator';
-import { buildSlo } from '../../../data/slo/slo';
+import { KibanaReactStorybookDecorator } from '../../../../utils/kibana_react.storybook_decorator';
+import { buildSlo } from '../../../../data/slo/slo';
 import { Overview as Component, Props } from './overview';
 
 export default {
