+++ x-pack/plugins/observability/public/utils/slo/convert_sli_apm_params_to_apm_app_deeplink_url.ts
@@ -6,7 +6,7 @@
  */
 
 interface Props {
-  duration: string;
+  duration?: string;
   environment: string;
   filter: string | undefined;
   service: string;
